pyproject.toml
[tool.poetry]
name = "googleapi"
version = "0.1.0"
description = ""
authors = ["scrbn72 <carlo.lombardi95@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-bigquery = "^3.22.0"
google-cloud-storage = "^2.16.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
